defect #997072 : remove uft job cleaner + build discarder in newly created jobs
